centralize resolveDataviewDatasetResource in package
