+++ packages/dataviews-client/src/index.ts
@@ -1,7 +1,9 @@
 export { default } from './dataviews-client'
 export {
   default as resolveDataviews,
+  resolveDataviewDatasetResource,
   UrlDataviewInstance,
   mergeWorkspaceUrlDataviewInstances,
 } from './resolve-dataviews'
+export { getGeneratorConfig, getDataviewsGeneratorConfigs } from './resolve-dataviews-generators'
 export { default as resolveEndpoint } from './resolve-endpoint'
